Move calculate_isoresolution_lines function out of slip_viewer
This means that we can use it from dials.export_bitmaps without implicitly
importing wx.
