+++ src/dials/util/export_bitmaps.py
@@ -15,11 +15,11 @@ from iotbx.detectors import FlexImage, FlexImage_d
 
 from dials.algorithms.image.threshold import DispersionThresholdDebug
 from dials.array_family import flex
+from dials.util.image_viewer import calculate_isoresolution_lines
 from dials.util.image_viewer.slip_viewer.flex_image import (
     get_flex_image,
     get_flex_image_multipanel,
 )
-from dials.util.image_viewer.spotfinder_frame import calculate_isoresolution_lines
 
 logger = logging.getLogger(__name__)
